Move all new/modified files from NewDesign folder to proper location
