+++ src/components/Gallery.vue
@@ -1,10 +1,9 @@
 <template>
-  <div v-if="imgs.length > 0" class="my-1">
+  <div v-if="imgs.length > 0" class="my-1 gallery">
     <v-card outlined :dark="dark" :color="color">
-      <v-card-title class="text-h4 font-weight-bold justify-center pa-0">{{
+      <v-card-title class="text-h4 font-weight-bold justify-center pa-0 mb-5">{{
         title
       }}</v-card-title>
-      <v-divider></v-divider>
       <v-card-text class="pa-0">
         <v-row no-gutters justify="center">
           <v-col
@@ -19,7 +18,7 @@
           >
             <v-hover v-slot="{ hover }">
               <v-img
-                class="clickable"
+                class="clickable image"
                 style="cursor: pointer"
                 :src="be + img.thumbnail"
                 aspect-ratio="1"
@@ -50,14 +49,15 @@
     </v-card>
     <v-dialog max-width="100%" v-model="dialog" content-class="elevation-0">
       <v-card color="rgba(0,0,0,0)" flat>
-        <v-carousel
+        <v-carousel align="center" justify="center"
+        class="carousel-content"
           hide-delimiters
           :show-arrows="imgs.length > 1"
           height="75vh"
           v-model="img_num"
         >
-          <v-carousel-item v-for="(img, i) in imgs" :key="i">
-            <v-img contain :src="be + img.image" aspect-ratio="1.7778"></v-img>
+          <v-carousel-item align="center" justify="center" v-for="(img, i) in imgs" :key="i">
+            <img class="gallery-image" :src="be + img.image" aspect-ratio="1.7778"/>
           </v-carousel-item>
         </v-carousel>
       </v-card>
@@ -105,3 +105,38 @@ export default class Gallery extends Vue {
   }
 }
 </script>
+
+<style scoped lang="scss">
+
+
+
+img.gallery-image{
+    margin: auto;
+    width: 100%;
+    max-width: 800px;
+    object-fit: contain;
+    height: 100%;
+}
+.gallery {
+  margin: 50px 0;
+
+
+
+  .v-image {
+    max-width: 300px;
+    margin: auto;
+
+  }
+  .image {
+    transition: filter 0.3s;
+    &:hover {
+      filter: brightness(0.5);
+    }
+  }
+
+  .theme--light.v-sheet--outlined {
+    background: transparent;
+    padding: 20px 0;
+  }
+}
+</style>
