+++ src/views/About.vue
@@ -1,13 +1,12 @@
 <template>
-  <div>
-    <v-card outlined class="ma-2">
-      <v-card-title
-        class="text-h4 justify-center font-weight-bold"
-        style="word-break: break-word"
-        >O Studenckim Kole Naukowym Informatyków "KOD"</v-card-title
-      >
-      <v-divider />
-      <v-card-text class="text-h6">
+  <div class="about-section">
+    <SectionHeader
+      title='O Studenckim Kole Naukowym Informatyków "KOD"'
+      subtitle="o nas"
+    >
+    </SectionHeader>
+    <div class="about-section-description">
+      <div class="about-section-description-inner">
         <p>
           Koło skupia głównie studentów kierunku Informatyka Wydziału
           Elektrotechniki i Informatyki i od początku działalności udaje się
@@ -33,9 +32,11 @@
           także udział w ogólnopolskich konferencjach dla młodych pracowników
           nauki i studentów.
         </p>
-      </v-card-text>
-    </v-card>
-    <p class="text-h4 text-center font-weight-bold">Historia koła</p>
+      </div>
+
+      <img :src="require('@/assets/about_1.jpg')" />
+    </div>
+    <SectionHeader title="Historia koła" subtitle="o nas"> </SectionHeader>
     <v-timeline class="ma-2" :dense="$vuetify.breakpoint.xs">
       <v-timeline-item
         fill-dot
@@ -43,6 +44,7 @@
         v-for="item in timelineItems"
         :key="item.timestamp"
         :icon="item.icon"
+        color="#55acee"
       >
         <template v-slot:opposite>
           <span
@@ -66,9 +68,15 @@
 
 <script lang="ts">
 import { Component } from 'vue-property-decorator';
+import SectionHeader from '../components//SectionHeader.vue';
+
 import Vue from 'vue';
 
-@Component
+@Component({
+  components: {
+    SectionHeader,
+  },
+})
 export default class Page404 extends Vue {
   private data() {
     return {
@@ -77,7 +85,8 @@ export default class Page404 extends Vue {
           timestamp: 1997,
           icon: 'mdi-handshake',
           title: 'Założenie Koła Naukowego',
-          text: 'Studenckie Koło Naukowe Informatyków "KOD" działa od 1997 r. \
+          text:
+            'Studenckie Koło Naukowe Informatyków "KOD" działa od 1997 r. \
             Założyli go studenci pierwszego roku uruchomionego wówczas kierunku Informatyka na Wydziale Elektrycznym Politechniki. \
             Opiekunem naukowym koła został dr inż. Andrzej Kubaszek (Zakład Podstaw Elektrotechniki i Informatyki)',
         },
@@ -85,7 +94,8 @@ export default class Page404 extends Vue {
           timestamp: 2003,
           icon: 'mdi-rename-box',
           title: 'Zmiana nazwy Koła',
-          text: 'W związku z rozwojem studenckiego ruchu naukowego nazwa Koła zmieniała się, najpierw na Koło Naukowe Informatyków – Programowanie Komputerów, a w 2003 r. \
+          text:
+            'W związku z rozwojem studenckiego ruchu naukowego nazwa Koła zmieniała się, najpierw na Koło Naukowe Informatyków – Programowanie Komputerów, a w 2003 r. \
                 Walne Zebranie Koła podjęło uchwałę o zmianie nazwy na Koło Naukowe Informatyków KOD. \
                 Ostatecznie nazwa została zmieniona na Studenckie Koło Naukowe Informatyków "KOD".',
         },
@@ -93,10 +103,61 @@ export default class Page404 extends Vue {
           timestamp: 2008,
           icon: 'mdi-account-tie',
           title: 'Zmiana opiekuna Koła',
-          text: 'Od 2008 r. funkcję opiekuna koła pełni dr inż. Bartosz Trybus (Katedra Informatyki i Automatyki)',
+          text:
+            'Od 2008 r. funkcję opiekuna koła pełni dr inż. Bartosz Trybus (Katedra Informatyki i Automatyki)',
         },
       ],
     };
   }
 }
 </script>
+
+<style scoped lang="scss">
+@use '@/styles/helpers' as *;
+@use '@/styles/components/buttons';
+.about-section {
+  margin-top: 50px;
+  @include responsiveLayout();
+  row-gap: 50px;
+  width: 100%;
+
+  .section-header {
+    text-align: center;
+    margin-bottom: 25px;
+  }
+
+  .v-sheet.v-card {
+    border-radius: 20px;
+  }
+
+  img {
+    width: 100%;
+    box-shadow: 0 0 8px 1px rgba(0, 0, 0, 0.35);
+    border-radius: 20px;
+  }
+  .about-section-description {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    margin-bottom: 50px;
+
+    .about-section-description-inner {
+      max-width: 90%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      flex-direction: column;
+    }
+  }
+
+  @media only screen and (max-width: 991px) {
+    .about-section-description {
+      grid-template-columns: 1fr;
+
+      .about-section-description-inner {
+        max-width: 100%;
+        margin-bottom: 25px;
+      }
+    }
+  }
+}
+</style>
