+++ src/views/Project.vue
@@ -1,98 +1,54 @@
 <template>
-  <div class="mt-2">
-    <v-row justify="center" v-if="project != undefined">
-      <v-col cols="12" sm="11" md="10" lg="9" xl="8">
-        <v-card tile elevation="0">
-          <v-card-title class="mb-2" style="word-break: break-word">
-            <p class="text-h2 text-left mb-0">{{ project.title }}</p>
-          </v-card-title>
-          <v-card-subtitle class="text-left pb-0">
-            <v-chip label>
-              <v-avatar left>
-                <v-icon>mdi-account-circle</v-icon>
-              </v-avatar>
-              <span
-                >{{ project.creator.first_name }}
-                {{ project.creator.last_name }}</span
-              >
-            </v-chip>
-            <p class="mb-0">
-              Data publikacji:
-              <b>{{
-                project.publication_date | moment('YYYY-MM-DD HH:mm:ss')
-              }}</b>
-            </p>
-          </v-card-subtitle>
-          <v-card-actions v-if="project.repository_link" class="py-0 ml-2">
-            <v-btn-cap
-              class="purple darken-2 white--text"
-              :href="project.repository_link"
-              target="_blank"
-            >
-              <v-icon left class>mdi-github</v-icon>
-              <span>Repozytorium</span>
-            </v-btn-cap>
-            <v-spacer />
-          </v-card-actions>
-          <v-card-text>
-            <markdown-it-vue
-              class="md-body text-left"
-              :content="project.text"
-              :options="markdownOptions"
-            />
-          </v-card-text>
-        </v-card>
-        <gallery breakpoints="xs6" :imgs="project.gallery" />
-      </v-col>
-    </v-row>
-    <v-row justify="center" v-if="project != undefined">
-      <v-col cols="12" sm="11" md="10" lg="9" xl="8" class="text-center">
-        <p>
-          Made with
-          <v-icon color="red">mdi-heart</v-icon> by
-        </p>
-        <div v-for="author in project.authors" :key="author.username">
-          <p>
-            <b
-              >{{ author.first_name }} "{{ author.username }}"
-              {{ author.last_name }}</b
-            >
-          </p>
-        </div>
-        <p>
-          at
-          <b>{{ project.section.name }}</b>
-        </p>
-      </v-col>
-    </v-row>
-    <v-speed-dial fixed right bottom direction="top" v-model="fab" v-if="role">
-      <template v-slot:activator>
-        <v-btn-cap
-          fab
-          v-model="fab"
-          class="text-body-1 font-weight-bold"
-          color="primary"
-        >
-          <v-icon>mdi-cog</v-icon>
-        </v-btn-cap>
-      </template>
-      <v-btn-cap
-        fab
-        color="orange"
-        :to="{ name: 'editProject', params: { id: $route.params.id } }"
-      >
-        <v-icon>mdi-pen</v-icon>
-      </v-btn-cap>
-      <v-btn-cap fab color="error" @click="dialog = true">
-        <v-icon>mdi-delete</v-icon>
-      </v-btn-cap>
-    </v-speed-dial>
+  <div class="project-page">
+    <section-header
+      :title="project.title"
+      :subtitle="project.section.name"
+      v-if="project"
+      class="project-title"
+    />
+
+    <div class="project-toolbar" v-if="isAdministrator">
+      <router-link :to="{ name: 'editProject', params: { id: $route.params.id } }" class="link">
+        <edit-icon class="icon" />
+        <span class="text">Edytuj projekt</span>
+      </router-link>
+
+      <span class="link" @click="deleteDialogOpen = true">
+        <trash-icon class="icon" />
+        <span class="text">Usuń projekt</span>
+      </span>
+    </div>
+
+    <div class="project-info" v-if="project">
+      <h2 class="authors-label" v-if="project.authors.length > 0">Autorzy</h2>
+      <div class="authors" v-if="project.authors.length > 0">
+        <v-chip label v-for="author in project.authors" :key="author.id">
+          <v-avatar left>
+            <v-icon>mdi-account-circle</v-icon>
+          </v-avatar>
+          <span>{{ author.first_name }} {{ author.last_name }}</span>
+        </v-chip>
+      </div>
+
+      <markdown-it-vue
+        class="project-description"
+        :content="project.text"
+        :options="markdownOptions"
+      />
+
+      <gallery breakpoints="xs6" :imgs="project.gallery" class="gallery" />
+    </div>
+
+    <div class="loading-wrapper" v-if="loading">
+      <v-progress-circular indeterminate color="primary" />
+    </div>
+
     <confirmation-dialog
-      v-if="dialog"
-      @yes="deleteProject($route.params.id)"
-      @no="dialog = false"
-      :text="dialogText"
-    ></confirmation-dialog>
+      v-if="deleteDialogOpen"
+      @yes="deleteProject"
+      @no="deleteDialogOpen = false"
+      text="Czy na pewno chcesz usunąć ten artykuł?"
+    />
   </div>
 </template>
 
@@ -100,66 +56,142 @@
 import { Component, Vue } from 'vue-property-decorator';
 import { ProjectsService } from '@/services/ProjectsService';
 import { ProjectModel } from '@/models/ProjectModel';
+import SectionHeader from '@/components/SectionHeader.vue';
+import EditIcon from '@/assets/icons/edit.svg?inline';
+import TrashIcon from '@/assets/icons/trash.svg?inline';
+import Gallery from '@/components/Gallery.vue';
 
-@Component
+@Component({
+  components: {
+    SectionHeader,
+    EditIcon,
+    TrashIcon,
+    Gallery,
+  },
+})
 export default class Project extends Vue {
-  private projectsService!: ProjectsService;
-  private project!: ProjectModel;
+  projectsService = new ProjectsService();
+  project: ProjectModel | null = null;
+  deleteDialogOpen = false;
 
-  private beforeCreate() {
-    this.projectsService = new ProjectsService();
+  markdownOptions = {
+    markdownIt: {
+      html: true,
+      linkify: true,
+    },
+    githubToc: {
+      anchorLink: false,
+    },
+  };
+
+  get loading() {
+    return !this.project;
   }
 
-  private mounted() {
-    this.projectsService
-      .getProject(+this.$route.params.id)
-      .then((project) => {
-        this.project = project;
-      })
-      .catch(() => {
-        this.$router.replace({ name: 'error404' });
-      });
+  async mounted() {
+    const projectId = Number.parseInt(this.$route.params.id, 10);
+
+    try {
+      const project = await this.projectsService.getProject(projectId);
+      this.project = project;
+    } catch (error) {
+      this.$router.replace({ name: 'error404' });
+    }
   }
 
-  get auth(): boolean {
-    return this.$store.getters.isAuthenticated;
-  }
-  get role(): boolean {
+  get isAdministrator(): boolean {
     return this.$store.getters.isAdministrator;
   }
-  get dialogText() {
-    return 'Czy na pewno chcesz usunąć artykuł "' + this.project.title + '"?';
-  }
 
-  private deleteProject(id: number) {
-    this.projectsService
-      .deleteProject(id)
-      .then((res) => {
-        if (res.status === 204) {
-          this.$store.dispatch('successMessage', 'Projekt został usunięty');
-          this.$router.replace({ name: 'projects', params: { page: '1' } });
-        }
-      })
-      .catch(() => {
+  async deleteProject() {
+    if (this.project) {
+      try {
+        await this.projectsService.deleteProject(this.project.id);
+        this.$store.dispatch('successMessage', 'Projekt został usunięty');
+        this.$router.replace({ name: 'projects', params: { page: '1' } });
+      } catch (error) {
         this.$store.dispatch('errorMessage', 'Błąd poczas usuwania projektu!');
-      });
-  }
-
-  private data() {
-    return {
-      project: this.project,
-      fab: false,
-      dialog: false,
-      markdownOptions: {
-        markdownIt: {
-          html: true,
-          linkify: true,
-        },
-        githubToc: {
-          anchorLink: false,
-        },
-      },
-    };
+      } finally {
+        this.deleteDialogOpen = false;
+      }
+    }
   }
 }
 </script>
+
+<style lang="scss" scoped>
+@use '@/styles/helpers' as *;
+
+.project-page {
+  @include responsiveLayout();
+
+  margin-top: 50px;
+  height: 100%;
+  width: 100%;
+
+  .project-title {
+    margin-bottom: 50px;
+  }
+
+  .loading-wrapper {
+    height: 300px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .project-toolbar {
+    background-color: $primary;
+    width: 100%;
+    margin-bottom: 50px;
+    border-radius: 10px;
+    padding: 10px 20px;
+    display: flex;
+    height: min-content;
+
+    & > * + * {
+      margin-left: 20px;
+    }
+
+    .link {
+      display: flex;
+      align-items: center;
+      cursor: pointer;
+      .icon {
+        width: 30px;
+        height: 30px;
+        color: $body-bg;
+        display: block;
+        margin-right: 10px;
+      }
+
+      .text {
+        font-weight: 600;
+        color: $body-bg;
+        text-decoration: none;
+        display: block;
+      }
+    }
+  }
+
+  .project-info {
+    .authors-label {
+      margin-bottom: 10px;
+    }
+
+    .authors {
+      display: flex;
+      flex-wrap: wrap;
+      margin-bottom: 50px;
+
+      & > * + * {
+        margin-left: 10px;
+      }
+    }
+
+    .project-description {
+      margin-bottom: 50px;
+    }
+  }
+}
+</style>
